backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai==2.24.0
google-generativeai==0.8.3
